Adding Arduino Sensor Program
This program contains code to implement logic determining the entry and exit (or false conditions) of a person. This is done by using state flags. Furthermore, the program also contains the encapsulation of the result of this logic in a LoRaWAN packet to be broadcasted and hopefully received by a LoRaWAN gateway.